backend/requirements.txt
pronotepy>=2.13.0
psycopg2-binary>=2.9.9
